add integration infra for disk failure prediction models
predict_device.py is a command line tool which receives a device serial
id (not the original textual UUID), and returns its failure prediction
result.

The tool prepares SMART metrics in a JSON format, which can be used by
any prediction model.

model.py is a placeholder for the actual model currently used in Ceph.

Sample input to the model is found in input_samples.

Signed-off-by: Yaarit Hatuka <yaarit@redhat.com>