server: do not post big reports to postgres
There are cases where cluster reports are very big due to their unpruned
crashes. Selecting such rows from public.report ends up in an error
(invalid memory alloc request size), and prevents importing new
reports to the cluster and crash tables.

If the report's size exceeds 100 MB, it will not be posted to postgres,
but it is still available as a file. We log such events to
/opt/telemetry/log/report_ep.log.

Signed-off-by: Yaarit Hatuka <yaarit@redhat.com>